requirements.txt
flask-restful==0.3.6
google-cloud-storage==1.11.0
sphinx==1.7.7
pylint==2.1.1